+++ src/app/app.module.ts
@@ -393,6 +393,8 @@ import { ModalDetailKnowledgeBaseComponent } from './knowledge-bases/modals/moda
 import { KnowledgeBaseTableComponent } from './knowledge-bases/modals/knowledge-base-table/knowledge-base-table.component';
 import { ModalErrorComponent } from './knowledge-bases/modals/modal-error/modal-error.component';
 import { KnowledgeBasesPreviousComponent } from './knowledge-bases-previous/knowledge-bases-previous.component';
+import { ModalUrlsKnowledgeBaseComponent } from './knowledge-bases/modals/modal-urls-knowledge-base/modal-urls-knowledge-base.component';
+import { AddContentMenuComponent } from './knowledge-bases/menu/add-content-menu/add-content-menu.component';
 import { UserModalComponent } from './users/user-modal/user-modal.component';
 import { ChatbotStatsModalComponent } from './bots/bots-list/chatbot-stats-modal/chatbot-stats-modal.component';
 
@@ -700,6 +702,8 @@ const appInitializerFn = (appConfig: AppConfigService, brandService: BrandServic
     KnowledgeBaseTableComponent,
     ModalErrorComponent,
     KnowledgeBasesPreviousComponent,
+    ModalUrlsKnowledgeBaseComponent,
+    AddContentMenuComponent,
     UserModalComponent,
     ChatbotStatsModalComponent
   ],